Bug fix for perustelut
A non-null perustelut value got multiplied to all applications which had null perustelut in the UI.
This is because react discards attributes with undefined or null values. value got discarded when
perustelut was null (_.get(hakemus, "arvio.perustelut", "") didn't help because it only works with undefined not null).
Discarding value in textarea resulted the textarea value to be not rendered at all and hence the previous value remained
in the UI.