Use same created/updated timestamp for all forms when copying avustushaku
This issue was uncovered by test that copies avustushaku and checks that
their timestamps match. If avustushaku was copied when minute changes
then the timestamps shown with minute granularity did not match on site.

The whole copy operation should be done in a single transactional
operation but that might be a bit bigger change.