Fix duration timestamps of the initial avustushaku
This migration fixes the invalid timestamp format of the initial
avustushaku from migration V1_1__Initial_data.sql. The invalid timestamp
break the maksatukset page unless you go and modify the avustushaku
which forces the timestamps to update to valid format.

The timestamps of the avustsuhaku have been modified to something else
in all environments which means WHERE condition is only true locally.