OPHY-371: Fix Fieldset styles that broke dropdown inputs in forms
The `.soresu-fieldset div` rule applied to all divs under `.soresu-fieldset`. This caused some divs inside the DropdownList from react-widgets to get `display: inline-block` and make them interactable even though they were not visible to the user. The style rule seems to have been unnecessary, as its purpose was probably to set the top-level inputs to `inline-block`, and to set a max-width to the whole fieldset.