Add approximate duration of fixture execution to timeout in hakemusTest
This reduces the need to have a long default timeout. It should also be not as necessary increase timeout per test after hitting timeout, and it should be easier to figure out a sane timeout. Initializing tests takes a long time (close to a minute), so it's nice to have some idea of how long each fixture should take defined right there in the fixture.