OPHY-411: Alert if email has not been sent in 60 minutes;
We currently use fluentbit to alert on-call user if there is an
error level log. Problem is the smtp server might be down once in
a while and now that we have a retry mechanism we dont want alerts
before the failed email has been retried for a while.

Also added error level logging to create-email-event as otherwise
now we would not get an alert if the storing failed for whatever
reason.