OPHY-523: Fix hakulomake lista components;
Problem was it called 2 times the callback which both in the end
called the redux store. When the first label edit changed
the formDraft, it caused a rerender and a new form draft into the
state. But because how react works the callback has a version of
the draft form that is not writable anymore when it calls
the value change form (after calling the label fn that cause the
rerender).

Tldr: fix is to do all changes in one callback that then in the
end updates the form in the store.