VA-301: warn if menoluokka needed fields are missing in form;
We had a problem were form was edited so that financing-plan was
removed and inside budget fields childrens ids were edited. This
made it so that menoluokka table was not updated and when clicking
menoluokat in arviointi we got null pointer exceptions as the
menoluokka db rows where different from the actual form.

Now we at least check that the form has valid fields for
menoluokka though the error message could be better.