OPHY-649-2: Shuffle managed dependencies around to hopefully be more meaningful
Clj-yaml is a transitive dependency of compojure-api, but we're not interested in clj-yaml, we want a recent enough snakeyaml. To achieve this, ring-middleware-format is upgraded to latest version which should be compatible with the abandonware package compojure-api, which upgrades clj-yaml to a new enough version that is compatible with snakeyaml 1.x. Since that version of clj-yaml doesn't specify a new enough snakeyaml (1.2x or so), it is forced to 1.33.