put scheduled reminder messages behind config
Our legacy clojure-spec tests do not support running scheduled jobs when
they start the server for tests. The jobs start running, and if the test
is quick enough it restarts and tries to clear the database, which in
practice destroys the state that the just started job is trying to run in.